rag-evaluation-backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-evaluation-backend"
version = "0.1.0"
description = "RAG Evaluation System Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.8.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.15.0",
    "python-jose[cryptography]>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "pandas>=2.2.0",
    "xlsxwriter>=3.2.0",
    "pyyaml>=6.0.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ipython>=8.12.0",
]

[project.scripts]
rag-eval = "app.main:app"

[tool.hatch.build.targets.wheel]
packages = ["app"]
